requirements.txt
<<<<<<< Updated upstream
torch>=1.12.0+cu116
torchvision>=0.13.0+cu116
Pillow>=9.2.0
PyQt5>=5.15.7
setuptools>=63.2.0
Django>=4.0.6
djangorestframework>=3.13.1
opencv-python>=4.6.0.66
numpy>=1.23.1
=======
torch>=1.12.0+cu116
torchvision>=0.13.0+cu116
Pillow>=9.2.0
PyQt5>=5.15.7
setuptools>=63.2.0
Django>=4.0.6
djangorestframework>=3.13.1
opencv-python>=4.6.0.66
numpy>=1.23.1
>>>>>>> Stashed changes
